No longer use the "user_catalog_table" storage option.
Now that CreateInitDecodingContext() accepts "need_full_snapshot" argument, we
no longer need to (mis)use the "user_catalog_table" storage option to protect
the table being squeezed against VACUUM.

While getting rid of the "user_catalog_table" option, also removed the code
that deals with "autovacuum_enabled". That wasn't very useful because the
window between disabling the autovacuum and starting the squeeze is usually
very short, so the chance that the setting it avoids concurrent autovacuum is
rather low. And since we never used any other storage options than these two,
the removal of "autovacuum_enabled" allows for further simplification of the
code.