Rework the background worker.
The previous versions of pg_squeeze used a single background worker that both
checked the schedule and bloat of the registered tables and executed the
squeeze_table() function. However, this way the execution of the
squeeze_table() function could have delayed the next schedule checks. To make
sure that no schedule is missed, the condition to check the scheduled time was

	<schedued time> <= now()

instead of

	EXTRACT(HOUR FROM <scheduled time>) = EXTRACT(HOUR FROM now())
	AND
	EXTRACT(MINUTE FROM <scheduled time>) = EXTRACT(MINUTE FROM now())

One particular problem of this approach was that table could have been
processed even hours after the schedule if it got bloated after the scheduled
time.

Now there are two background workers: 1) "scheduler worker" which only checks
the schedule and thus should not miss any minute, 2) "squeeze worker" which
checks if the tables scheduled are really bloated, and if so, it squeezes
them.

This commit bumps the version number because it introduces schema changes.