Reduce disk consumption during the initial load.
Now we regularly decode the WAL produced during the initial load so that the
replication slot can advance during the load. Although we cannot apply the
decoded changes until the initial load has completed, the amount of data to
keep is relatively small because what we decode at this stage is mostly the
changes produced by the initial load itself, and these changes should be
ignored. We only need to keep the changes done by other transactions.

In order to avoid decoding of the data changes generated by the initial load,
we set their "replication origin" to specific value, and use output plugin
callback to filter such data out.