Reworked the interface to the core functionality.
Due to the commit 240e0dbacd in the PG core, it's no longer possible to use
the original squeeze_table() function because this way the FMGR registers a
snapshot automatically, which causes an ERROR. Therefore, the "squeeze worker"
was redesigned so it calls the corresponding C function directly.

In order to keep the squeeze_table() function working at least for testing and
troubleshooting, shared memory infrastructure is added here which allows a
regular backend to submit a task and wait until the "squeeze worker" finishes
it.

That shared memory infrastructure is now also used to check that only a single
worker exists per database. We no longer use the "heavy-weight" locks because
these should only be used in a transaction, and because a transaction abort
releases them. One particular consequence was that the stop_worker() function
used to miss some workers (because it used to retrieve the worker PID from the
`pg_locks` view).