Re-introduced arguments of the squeeze_table() function.
I removed them from v1.6 because "ind_tablespaces" is varlena and it's not
trivial to pass it to the worker via shared memory. However some users might
still appreciate the original signature. Eventually I decided to allocate a
fixed-size area for the array which should be sufficient in normal use cases.