Handle non-existing table name.
In the "interactive" mode (i.e. when the squeeze_table() function is used to
create the task), an incorrect table name caused the squeeze worker to
exit. The correct behavior is to insert a record into the "squeeze.errors"
table and go on.

Refactored the error handling code while working on this.