Remove an unnecessary call of start_workers() from regression test.
An earlier commit teaches the squeeze_table() function to start a squeeze
worker if one is not already running for given database. Thus we don't need to
do so explicitly in the test script.

Moreover, there seeme to be race conditions causing an error: 1) the squeeze
worker is started explicitly by start_worker(), 2) squeeze_table() doesn't see
the worker's "shared memory slot" initialized yet, so it also tries to start
the worker, 3) the worker launched by start_worker() completes the startup, 3)
the worker launched by squeeze_table() sees the first worker, so it exits
immediately, 5) the squeeze_table() function receives the BGWH_STOPPED code
from WaitForBackgroundWorkerStartup(), which is not expected, so it raises
ERROR.