Only start the squeeze worker if there is some work to do.
Instead of letting one or more squeeze workers per database run all the time,
it seems better to let the scheduler worker start them only if there are
tables to squeeze. This design also simplifies the squeeze_table() function:
it's simpler to start a new worker and wait until it exits than to communicate
with an existing worker via shared memory.