Move setup of the replication slots away from the squeeze worker.
During the setup of the snapshot builder for logical decoding, the worker may
need to wait until the currently running transactions complete. However some
of those transactions may be those of the "initial load" being performed by
other workers. This can effectively prevent the squeeze workers from
processing multiple tables at a time.

The new design is such that the snapshot builder setup is performed by the
scheduler worker and both the slot and the snapshot is passed to each worker
via shared memory before the workers are started.